mutli-agent-consensus-main/requirements.txt
numpy>=1.20.0
matplotlib>=3.5.0
pandas>=1.3.0
openai>=1.0.0
statsmodels>=0.13.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio

# Configuration and plotting
pyyaml>=6.0
seaborn>=0.13.0

# Tokenization diagnostics
tiktoken>=0.7.0

# Optional provider SDKs used by this repo
google-generativeai>=0.8.0

# PyTorch and ML dependencies for open-source LLM experiments
torch==2.8.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.56.2
accelerate==1.10.1
bitsandbytes==0.47.0
sentencepiece==0.2.1
safetensors==0.6.2
huggingface-hub>=0.34.0

# Optional GGUF backend
llama-cpp-python>=0.3.16
